feat: implement snapshot voting at proposal start height
Change applies to both core and action proposals. One caveat is we cannot use the read-only function from the trait so we have to hardcode the contract, but this pattern may be preferable since we are using templates on the backend anyway.

To put it another way, instead of initializing these contracts with the token contract, we can reference it directly since it gets deployed first. This should simplify the scripts / info needed to call it as well.