fix: hardcode token contract instead of init
This simplifies the code we have to test per dao, and we can insert this contract reference using the templates when generating, so the init was an extra step that wasn't needed. Every dao is focused around one token, so we can directly reference it and let other values change, like which treasury is compared for voting.