feat: model change to simplify buy/sell
This uses the read-only functions get-in and get-out to calculate the values instead of doing the same math again within the function. It ensures the read-only functions are accurate by using them to do the same calculation within the public function. This would require adding fields to the return values of get-in and get-out to provide all parameters necessary for buy/sell.