refactor: integrate faktory as token and dex
Instead of keeping separate contracts and duplicating efforts, we can make the faktory format the defaults for the token and dex since that is what will be used on deploys moving forward. We have git history and on-chain data if we need to access older contracts, or can bring them in as add-ons when testing out ideas.