fix: add more output to test
Not clear why this fails, block height is measured consistently throughout but the value that gets stored in the contract is different. Will try to reproduce on testnet.

Also noted that while we're in Epoch 3 with the contracts the clarity_version is still set to 2. This works and deploys (should it?) however switching to clarity_version 3 and stacks-block-height causes errors with all get-stacks-block-info? functions. Rock and a hard place.