refactor: remove constant use for contract-call operations
This removes the constant and instead puts it into a comment following the line where the contract-call is. We'll likely have to map this to the it.key format in the backend but will at least clear the errors for now so we can settle on a format.

Still seems worth organizing as much of the template info at the top as possible.