BREAKING: update end block to indicate when execution can happen
When a proposal is created it starts with a voting delay, voting period, execution delay, execution period, then expired depending on when it is concluded. This flips the logic for endBlock to include the execution delay so it's easier to react to concluding the proposal based on the block height rather than having to calculate it.