refactor: remove unusable core proposals
Creates a test that runs all core proposals in succession to make sure all is working, detected a few where switching context doesn't work as expected, on less relevant items like using a proposal to pay an invoice or make a deposit.

If we end up needing that type of functionality there should be some ways we can work magic later.