fix: match traits string vs contract name for excluded check
This is really a band-aid to make sure the check looks at all other contracts files, it excludes the trait directory by default for the dao but the smart wallet trait lives next to the file. Probably some reorg of the files can fix this but for now the test will just exclude matching a test file for any contract name that includes "traits".